UserServlet and API updated, LogoutServlet created.

+++ src/main/java/com/example/cinema_test/controller/api/ShowApi.java
@@ -1,13 +1,15 @@
 package com.example.cinema_test.controller.api;
 
 import com.example.cinema_test.controller.exception.ExceptionWrapper;
+import com.example.cinema_test.model.entity.Cinema;
+import com.example.cinema_test.model.entity.Show;
+import com.example.cinema_test.model.service.CinemaService;
 import com.example.cinema_test.model.service.ShowService;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import lombok.extern.slf4j.Slf4j;
-
 
 
 @Slf4j
@@ -17,14 +19,29 @@ public class ShowApi {
     @Inject
     private ShowService showService;
 
+    @Inject
+    private CinemaService cinemaService;
+
     @DELETE
     @Path("/{id}")
     public Response delete(@PathParam("id") Long id) {
         try {
+            Show showToRemove = showService.findById(id);
+
+            if (showToRemove.isAvailable()) {
+                return Response.status(Response.Status.NOT_ACCEPTABLE)
+                        .entity("Show is active on a cinema !!!")
+                        .build();
+            } else {
                 showService.remove(id);
+                for (Cinema cinema : cinemaService.findAll()) {
+                    cinema.getShowList().removeIf(show -> show.getId().equals(showToRemove.getId()));
+                    cinemaService.edit(cinema);
+                }
                 log.info("Show removed successfully-ID : " + id);
                 return Response.accepted().build();
-        }catch (Exception e) {
+            }
+        } catch (Exception e) {
             log.error(ExceptionWrapper.getMessage(e).toString());
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
                     .entity("An error occurred: " + e.getMessage())
@@ -48,7 +65,7 @@ public class ShowApi {
                         .entity("No records found for name: " + name)
                         .build();
             }
-        }catch (Exception e) {
+        } catch (Exception e) {
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
                     .entity("An error occurred: " + e.getMessage())
                     .build();
